requirements.txt
streamlit>=1.28.0
chromadb>=0.4.15
sentence-transformers>=2.2.2
pymupdf>=1.23.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
ollama>=0.1.7
pandas>=2.0.0
numpy>=1.24.0